create a login/logout routes moved the login form to a component login logout redirect correctly
